split utils.py
