+++ quri_algo/algo/compiler/base_classes.py
@@ -29,7 +29,7 @@ from quri_algo.algo.interface import (
     QuantumAlgorithmResult,
     VariationalAlgorithmResultMixin,
 )
-from quri_algo.algo.utils import timer
+from quri_algo.algo.utils.timer import timer
 from quri_algo.circuit.interface import CircuitFactory
 from quri_algo.core.cost_functions.base_classes import CostFunction
