+++ quri_algo/algo/compiler/qaqc.py
@@ -19,7 +19,8 @@ from quri_algo.algo.compiler.base_classes import (
     QuantumCompilerGeneric,
 )
 from quri_algo.algo.interface import Analysis, Analyzer, LoweringLevel
-from quri_algo.algo.utils import CircuitMapping, timer
+from quri_algo.algo.utils.mappings import CircuitMapping
+from quri_algo.algo.utils.timer import timer
 from quri_algo.circuit.interface import CircuitFactory
 from quri_algo.core.cost_functions.base_classes import CostFunction
 from quri_algo.core.cost_functions.utils import prepare_circuit_hilbert_schmidt_test
