+++ quri_algo/algo/interface.py
@@ -16,7 +16,8 @@ from quri_parts.algo.optimizer import OptimizerState
 from quri_parts.backend.units import TimeValue
 from quri_parts.circuit import NonParametricQuantumCircuit
 
-from quri_algo.algo.utils import CircuitMapping, timer
+from quri_algo.algo.utils.mappings import CircuitMapping
+from quri_algo.algo.utils.timer import timer
 
 LoweringLevel = Enum(
     "LoweringLevel",
